Move TOTP related methods to UserTotpMethods concern
